+++ src/Service/Location/Find.php
@@ -3,7 +3,7 @@
   * Find.php
   * Description: Service Location
   * @Author : M.V.M.
-  * @Version 1.0.5
+  * @Version 1.0.18
 **/
 declare(strict_types=1);
 
@@ -27,8 +27,7 @@ use App\Service\BaseFind;
       $query         = (array) $request->getQueryParams();
       $tableClass    = new Location($parameters->getPrefix(),$body);                 // (1) Modify in the new class the tableClass, if the class is copied
       $find          = new BaseFind($tableClass,$parameters);
-      $result        = (array) $find->getAll($query);
-      return $result;
+      return (array) $find->getAll($query);
     }
 
     public function getOne(Request $request,array $args, BaseParameters $parameters)
@@ -36,9 +35,6 @@ use App\Service\BaseFind;
       $body         = (array) $request->getParsedBody();
       $tableClass   = new Location($parameters->getPrefix(),$body);                // (1) Modify in the new class the tableClass, if the class is copied
       $find         = new BaseFind($tableClass,$parameters);
-      $result       = (array) $find->getOne($args);
-      return $result;
+      return (array) $find->getOne($args);
     }
-
 }
-