+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.0.18 (1013-12-28)
+
+* New class mailer control to login, controller and auth.
+## 1.0.17 (2023-12-22)
+
+* Fixed variable bug in templates in FrondEnd.
+
 ## 1.0.16 (2023-12-18)
 
 * Simplify php code in frontend templates